pyproject.toml
[tool.poetry]
name = "industrial-logistics-platform"
version = "1.0.0"
description = "Industrial Logistics Data Platform with dbt, Data Quality and Data Governance"
authors = ["Romilson Luis <romilsonlonan@gmail.com>"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.0.0"
psycopg2-binary = "^2.9.0"
great-expectations = "^0.18.0"
pytest = "^8.0.0"
python-dotenv = "^1.0.0"
prometheus-client = "^0.19.0"
pyyaml = "^6.0.0"

[tool.poetry.dev-dependencies]
black = "^24.0.0"
ruff = "^0.2.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]